requirements.txt
flask
flask_session
boto3
python-dotenv